tsql-core/src/test/java/com/xy/xsql/tsql/converter/functions/mathematical/ATan2ConverterTest.java
package com.xy.xsql.tsql.converter.functions.mathematical;

import com.xy.xsql.block.core.printer.MetaBlockPrinter;
import com.xy.xsql.block.model.BlockMeta;
import org.junit.Assert;
import org.junit.Test;

import java.io.StringWriter;

import static org.junit.Assert.*;

/**
 * Created by xiaoyao9184 on 2018/5/31.
 */
public class ATan2ConverterTest {

    @Test
    public void testMetaPrint() throws Exception {
        BlockMeta b = ATan2Converter.meta;

        StringWriter writer = MetaBlockPrinter.print(b);

        System.out.println(writer);
        Assert.assertEquals(writer.toString(),
                "<ATN2> ::=\n" +
                        "ATN2 ( float_expression , float_expression2 )");
    }

}